refactored component exports and imports to utilize components/index file
