+++ client/components/index.js
@@ -3,6 +3,12 @@
  * This way, we can import all of our components from the same place, rather than
  * having to figure out which file they belong to!
  */
-export {default as Navbar} from './navbar'
-export {default as UserHome} from './user-home'
-export {Login, Signup} from './auth-form'
+export {default as Navbar} from './Navbar'
+export {default as UserHome} from './UserHome'
+export {Login, Signup} from './AuthForm'
+export {default as AddGame} from './AddGame'
+export {default as JoinGame} from './JoinGame'
+export {default as CameraCanvas} from './CameraCanvas'
+export {default as Opentok} from './Opentok'
+export {default as PartnerMode} from './PartnerMode'
+export {default as SoloMode} from './SoloMode'
